Restructure and rename register/login pages
