+++ src/containers/Login/Login.css
@@ -1,4 +1,4 @@
-.Main{
+.Login{
   font-family: helvetica;
   width:80%;
   margin-left: 10%;
