+++ src/containers/Register/Register.css
@@ -1,4 +1,4 @@
-.Main{
+.Register{
   font-family: helvetica;
   width:80%;
   margin-left: 10%;
