--- src/containers/RegisterContainer/RegisterContainer.css
@@ -1,6 +1,0 @@
-.Main{
-  width:30%;
-  margin-left: 35%;
-  margin-right:35%;
-  background-color: white;
-}
